+++ spec/spec_helper.rb
@@ -72,7 +72,7 @@ RSpec.configure do |config|
         self.use_env          = false
         self.knockout_prefix  = nil
         self.overwrite_arrays = true
-        self.schema           = nil if RUBY_VERSION >= '2.1'
+        self.schema           = nil
         class_variable_set(:@@_ran_once, false)
       end
     end